+++ nata-monkey/src/main/java/com/nata/action/ActionType.java
@@ -8,6 +8,8 @@ package com.nata.action;
 public class ActionType {
     public static final String UNKNOWN= "unknown";
     public static final String TAP = "tap";
+    public static final String LONGCLICK= "longclick";
+
     public static final String SWIPE= "swipe";
     public static final String INPUT= "input";
     public static final String BACK= "back";