+++ nata-monkey/src/main/java/com/nata/action/LongClickAction.java
@@ -6,26 +6,27 @@ import com.nata.element.Element;
 /**
  * Author: Calvin Meng
  * Blog: mclspace.com  Email: rdmclin2@gamil.com
- * Update: 2016-01-13 16:48
+ * Update: 2016-03-24 21:23
  */
-public class TapAction extends Action{
+public class LongClickAction extends Action{
+
     private AdbDevice device = null;
     private Element element = null;
-    public TapAction(AdbDevice device,Element element){
-        super.setName(ActionType.TAP);
+    public LongClickAction(AdbDevice device,Element element){
+        super.setName(ActionType.LONGCLICK);
         this.device = device;
         this.element = element;
     }
 
     @Override
     public void fire() {
-        device.tap(element.getX(),element.getY());
+        device.longPress(element.getX(),element.getY());
     }
 
 
     @Override
     public String toString() {
-        return "TapAction{" +
+        return "LongClickAction{" +
                 "element=" + element +
                 '}';
     }