+++ nata-monkey/src/main/java/com/nata/action/StartAppAction.java
@@ -18,5 +18,6 @@ public class StartAppAction extends Action{
     @Override
     public void fire() {
         device.startActivity(pkgAct);
+        device.sleep(2000);
     }
 }