+++ nata-monkey/src/main/java/com/nata/monkeys/AbstractMonkey.java
@@ -4,7 +4,9 @@ import com.nata.cmd.AdbDevice;
 import com.nata.element.DumpService;
 import com.nata.element.Element;
 import com.nata.element.UINode;
+import com.nata.state.State;
 import org.dom4j.DocumentException;
+import org.xml.sax.SAXException;
 
 import java.io.*;
 import java.util.ArrayList;
@@ -16,45 +18,70 @@ import java.util.List;
  * Update: 2016-01-13 13:59
  */
 public abstract class AbstractMonkey {
-    private final String name ;
-    private final String pkg ;
-    private final String act ;
+    private final String name;
+    private final String pkg;
+    private final String act;
     private final AdbDevice device;
 
-
-    public AbstractMonkey(String name,String pkg,String act,AdbDevice device){
+    public AbstractMonkey(String name, String pkg, String act, AdbDevice device) {
         this.name = name;
         this.pkg = pkg;
         this.act = act;
         this.device = device;
     }
 
-    public String getPkgAct(){
-        String pkgAct = this.getPkg()+"/"+ this.getAct();
+    public String getPkgAct() {
+        String pkgAct = this.getPkg() + "/" + this.getAct();
         return pkgAct;
     }
 
-    public boolean isInCurrentPkg(){
+
+    public State getCurrentState() {
+        String curActivity = getCurrentActivity();
+        State state = new State(curActivity);
+        List<UINode> uiNodes = GrabCurrentUi();
+
+
+        // Add the uinodes with resourceId;
+        for(UINode node : uiNodes){
+            if(!node.getResourceId().equals("")){
+               state.addUINode(node);
+            }
+        }
+
+        return state;
+    }
+
+    public String getCurrentActivity(){
+        return device.getCurrentActivity();
+    }
+
+
+    public boolean isInCurrentPkg() {
         return device.getCurrentPackageName().equals(pkg);
     }
 
-    protected AdbDevice getDevice(){
+    protected AdbDevice getDevice() {
         return device;
     }
 
 
-    protected List<UINode>  GrabCurrentUi(){
+    protected List<UINode> GrabCurrentUi() {
         //Get dump file
         File dumpFile = device.dumpUI();
         try {
             List<UINode> list = DumpService.getNodes(dumpFile);
-
-            for (UINode node: list) {
+            for (UINode node : list) {
                 System.out.println(node);
             }
             return list;
-
         } catch (DocumentException e) {
+            e.printStackTrace();
+        } catch (UnsupportedEncodingException e) {
+            e.printStackTrace();
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+        } catch (SAXException e) {
             e.printStackTrace();
         }
 
@@ -62,15 +89,15 @@ public abstract class AbstractMonkey {
     }
 
     //TODO: detect errors,how?
-    public boolean isCrashed(){
+    public boolean isCrashed() {
         return false;
     }
 
-    public void Tap(Element element){
-        device.tap(element.getX(),element.getY());
+    public void Tap(Element element) {
+        device.tap(element.getX(), element.getY());
     }
 
-    public void SleepShort(){
+    public void SleepShort() {
         device.sleep(500);
     }