+++ nata-monkey/src/main/java/com/nata/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Author: Calvin Meng
+ * Blog: mclspace.com  Email: rdmclin2@gamil.com
+ * Update: 2016-03-24 15:16
+ */
+package com.nata;